--- Publish.bat
@@ -1,3 +1,0 @@
-@echo off
-powershell.exe -NoProfile -ExecutionPolicy Bypass -File "%~dp0Publish.ps1"
-timeout 30
