--- TuneLift.csproj
@@ -1,38 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-	<NoWarn>$(NoWarn);CA1416</NoWarn>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-
-	<!-- Assembly information -->
-    <ApplicationIcon>TuneLift.ico</ApplicationIcon>
-	<Authors>Richard Lawrence</Authors>
-	<Product>TuneLift</Product>
-	<Description>Export iTunes audio playlists as standard or extended .m3u files</Description>
-	<AssemblyTitle>Export iTunes audio playlists as standard or extended .m3u files</AssemblyTitle>  
-	<PackageProjectUrl>https://github.com/mrsilver76/tunelift</PackageProjectUrl>
-	<RepositoryUrl>https://github.com/mrsilver76/tunelift</RepositoryUrl>
-	<Copyright>Copyright © 2020-2025 Richard Lawrence</Copyright>
-	<GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-
-	<!-- Version information -->
-	<!-- Remember major.minor.BUILD.revision! -->
-    <Version>1.0.0.1</Version>
-    <FileVersion>1.0.0.1</FileVersion>
-    <AssemblyVersion>1.0.0.1</AssemblyVersion>
-    <AnalysisLevel>latest-recommended</AnalysisLevel>
-    <EnableNETAnalyzers>True</EnableNETAnalyzers>
-	</PropertyGroup>
-
-  <ItemGroup>
-    <Content Include="TuneLift.ico" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="ini-parser" Version="2.5.2" />
-  </ItemGroup>
-
-</Project>
