+++ TuneLift/Publish.ps1
@@ -1,10 +1,10 @@
 # PowerShell script to build and package a C# project for multiple architectures
-# Version 1.0.2 - 4th June 2025
+# Version 1.0.3 - 7th August 2025
 
 # === User Configurable Section ===
 
 # Hardcoded list of target architectures
-#$architectures = @("win-x64", "linux-x64", "linux-arm64", "osx-arm64", "osx-x64")
+#$architectures = @("win-x64", "linux-x64", "linux-arm", "linux-arm64", "osx-arm64", "osx-x64")
 $architectures = @("win-x64")
 
 # === Helper Functions ===
@@ -38,8 +38,10 @@ function Get-SemanticVersion {
     $revision = $parts[3]
 
     $semVer = "$major.$minor.$revision"
+	# If the build number is > 0 then this is an interim release between
+	# the last one and the next one, so append "-buildX" to the version
     if ([int]$build -gt 0) {
-        $semVer += "-pre$build"
+        $semVer += "-build$build"
     }
     return $semVer
 }
