+++ TuneLift/TextUtils.cs
