+++ TuneLift/TuneLift.csproj
@@ -20,9 +20,11 @@
 
 	<!-- Version information -->
 	<!-- Remember major.minor.BUILD.revision! -->
-    <Version>1.0.0.0</Version>
-    <FileVersion>1.0.0.0</FileVersion>
-    <AssemblyVersion>1.0.0.0</AssemblyVersion>
+    <Version>1.1.0.0</Version>
+    <FileVersion>1.1.0.0</FileVersion>
+    <AssemblyVersion>1.1.0.0</AssemblyVersion>
+    <AnalysisLevel>latest-recommended</AnalysisLevel>
+    <EnableNETAnalyzers>True</EnableNETAnalyzers>
 	</PropertyGroup>
 
   <ItemGroup>
